added confirmation function
